requirements.txt
apache-airflow==2.10.4
pandas==2.3.1
requests==2.32.4
boto3==1.40.0
psycopg2-binary==2.9.10
pydantic==2.11.7
